breaking: componentIndex changed to byte in CommandMessage and RpcMessage (#2963)

+++ Assets/Mirror/Runtime/Messages.cs
@@ -27,7 +27,7 @@ namespace Mirror
     public struct CommandMessage : NetworkMessage
     {
         public uint netId;
-        public int componentIndex;
+        public byte componentIndex;
         public int functionHash;
         // the parameters for the Cmd function
         // -> ArraySegment to avoid unnecessary allocations
@@ -37,7 +37,7 @@ namespace Mirror
     public struct RpcMessage : NetworkMessage
     {
         public uint netId;
-        public int componentIndex;
+        public byte componentIndex;
         public int functionHash;
         // the parameters for the Cmd function
         // -> ArraySegment to avoid unnecessary allocations
